model-orchestrator/requirements.txt
Flask==2.0.2
kubernetes==23.3.0
urllib3==1.26.8
setuptools==57.0.0
PyYAML==6.0